move functions to Common.py
