+++ Python/Common.py
@@ -1,5 +1,5 @@
 
-
+# 1 - yn_input_check: asks for a y/n input (case insensitive) by the user. Anything else is not accepted.
 def yn_input_check():
 
     char_input = ""
@@ -19,3 +19,28 @@ def yn_input_check():
         else:
 
             print("\nThe inserted value is not valid, please input \"y\" or \"n\"\n")
+
+#2 - int_input_check: asks for an integer by the user, then returns that value. Anything else is not accepted.
+def int_input_check():
+
+    int_input = -1
+
+    while (int_input < 0):
+
+        try:
+            
+            int_input = int(input())
+    
+            while(int_input < 0):
+
+                print("\nThe inserted value is not valid, please input a number higher than 0:\n")
+
+                int_input = int(input())
+
+            return int_input
+        
+        except:
+
+            print("\nThe inserted value is not valid, please input an integer number higher than 0:\n")
+
+            int_input = -1
